Made a solid frame
Exported the geometry as .STEP, imported with the option to merge/knit surfaces to a solid, healed an error at the left ear, made the solid hollow with 30mm thickness, then offset the panel perimeters by 30mm and did a cut extrude to get the frame.  The latest is AE5i_30mmShellFrame.  Made a new file because this causes the geometry to lose associativity to the construction geometry.